Added display options to the high-level API and re-factoring
This change included re-factoring the local API to use strings rather than enums. The high-level API does use enums. New methods have been added to the enums to allow easy transitioning between the raw values and the Java objects.

enun.toRaw() will provide the correct low-level string. This should be used over any other method of converting an enum to a string. A synonym for this is available in ApiValue.raw(enum). All of the enum types in the high-level API implement the ApiValue interface.

EnumType.toEnum(String) will convert the raw value to the enum. This method will not throw any exceptions and is null safe. If a null value is given, a null value will be returned.