Add high-level API methods to activate and operate on widgets
Widgets are instances of applications. When you operate on an app, the library assumes you want to work with the first widget (first instance). For many users and apps this will be simpler and easier. 

However, apps can be duplicated (even some of the core apps like weather). When the device has more than one instance of a given app, you can explicitly refer to an instance by using the appropriate widget.